Revert "Merge pull request #6617 from geoberle/gather-observability-azure-monitor"
This reverts commit 2126eb280b5b117fbcb060be38ebe876e95402bc, reversing
changes made to 10c24e68ffe8a4c8b20eb23e2ca82618d23f3b22.
